# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/layer-groups": {
      "get": {
        "operationId": "listLayerGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      },
      "post": {
        "operationId": "createLayerGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLayerGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      }
    },
    "/api/v1/admin/layers": {
      "get": {
        "operationId": "listAdminLayers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      },
      "post": {
        "operationId": "createLayer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLayerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      }
    },
    "/api/v1/admin/revisions/{revisionId}": {
      "get": {
        "operationId": "getRevision",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      }
    },
    "/api/v1/admin/revisions/{revisionId}/workspace": {
      "get": {
        "operationId": "getRevisionWorkspace",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      }
    },
    "/api/v1/admin/revisions/{revisionId}/features": {
      "get": {
        "operationId": "listAdminFeatures",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bbox",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      },
      "post": {
        "operationId": "createFeature",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureMutationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      }
    },
    "/api/v1/admin/revisions/{revisionId}/features/{featureId}": {
      "patch": {
        "operationId": "updateFeature",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featureId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeatureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      },
      "delete": {
        "operationId": "deleteFeature",
        "parameters": [
          {
            "name": "revisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featureId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "requestId"
                      ],
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "",
        "tags": [
          "admin-layers"
        ]
      }
    },
...
  },
  "info": {
    "title": "DanangMap API",
    "description": "DanangMap v2 spatial CMS contract",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-danangmap_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
...
      "CreateLayerGroupDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "government",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "title": {
            "type": "string",
            "example": "Cơ quan hành chính",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "displayOrder": {
            "type": "integer",
            "default": 0
          },
          "defaultVisible": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "slug",
          "title"
        ]
      },
      "LayerFieldDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "address",
            "pattern": "^[a-z][a-z0-9_]{1,63}$"
          },
          "label": {
            "type": "string",
            "example": "Địa chỉ",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "long_text",
              "number",
              "integer",
              "boolean",
              "date",
              "datetime",
              "url",
              "email",
              "phone",
              "enum",
              "multi_enum",
              "address",
              "image",
              "attachment"
            ]
          },
          "icon": {
            "type": "string",
            "example": "map-pin",
            "maxLength": 64
          },
          "required": {
            "type": "boolean",
            "default": false
          },
          "public": {
            "type": "boolean",
            "default": true
          },
          "searchable": {
            "type": "boolean",
            "default": false
          },
          "filterable": {
            "type": "boolean",
            "default": false
          },
          "sortable": {
            "type": "boolean",
            "default": false
          },
          "sensitive": {
            "type": "boolean",
            "default": false
          },
          "offlineCache": {
            "type": "boolean",
            "default": true
          },
          "defaultValue": {
            "nullable": true,
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {}
              }
            ]
          },
          "validation": {
            "type": "object",
            "additionalProperties": true,
            "default": {}
          },
          "options": {
            "type": "array",
            "items": {},
            "maxItems": 100,
            "default": []
          },
          "displayOrder": {
            "type": "integer",
            "default": 0
          }
        },
        "required": [
          "key",
          "label",
          "type"
        ]
      },
      "CreateLayerDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "administrative-offices",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "displayOrder": {
            "type": "integer",
            "default": 0
          },
          "title": {
            "type": "string",
            "example": "Trụ sở hành chính",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "geometryMode": {
            "type": "string",
            "enum": [
              "point",
              "circle",
              "polyline",
              "polygon",
              "mixed"
            ]
          },
          "allowedGeometryKinds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 7,
            "items": {
              "type": "string",
              "enum": [
                "point",
                "multipoint",
                "line",
                "multiline",
                "polygon",
                "multipolygon",
                "circle"
              ]
            }
          },
          "fields": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LayerFieldDto"
            }
          },
          "style": {
            "type": "object",
            "additionalProperties": true,
            "default": {}
          },
          "renderConfig": {
            "type": "object",
            "additionalProperties": true,
            "default": {}
          },
          "popupConfig": {
            "type": "object",
            "additionalProperties": true,
            "default": {}
          }
        },
        "required": [
          "slug",
          "title",
          "geometryMode",
          "allowedGeometryKinds",
          "fields"
        ]
      },
      "FeatureMutationDto": {
        "type": "object",
        "properties": {
          "geometry": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "type": "Point",
              "coordinates": [
                108.2208,
                16.0678
              ]
            }
          },
          "geometryKind": {
            "type": "string",
            "enum": [
              "point",
              "multipoint",
              "line",
              "multiline",
              "polygon",
              "multipolygon",
              "circle"
            ]
          },
          "radiusM": {
            "type": "number",
            "nullable": true,
            "minimum": 0.001,
            "maximum": 10000000
          },
          "externalSource": {
            "type": "string",
            "maxLength": 200
          },
          "externalId": {
            "type": "string",
            "maxLength": 500
          },
          "properties": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "geometry",
          "geometryKind",
          "properties"
        ]
      },
      "UpdateFeatureDto": {
        "type": "object",
        "properties": {
          "geometry": {
            "type": "object",
            "additionalProperties": true
          },
          "geometryKind": {
            "type": "string",
            "enum": [
              "point",
              "multipoint",
              "line",
              "multiline",
              "polygon",
              "multipolygon",
              "circle"
            ]
          },
          "radiusM": {
            "type": "number",
            "nullable": true,
            "minimum": 0.001,
            "maximum": 10000000
          },
          "properties": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}